Database cleanup tasks improvements
- Fix bug where dangling criterion records were left behind after commands are deleted
- Move all bulk deletion APIs for entities to READ_COMMITTED transactions to reduce likelyhood of db locks impacting performance with default REPEATABLE_READ
- Make all bulk deletion APIs use a select ids -> delete by ids pattern that leverages deleting by entities
	- While this is slightly less ideal than just outright deleting it does let us leverage hibernate to delete relational records like criterion without us having to do it
- Reduce complexity on job deletion logic to just delete in batches instead of batches of batches
- Add status filter to job deletion logic so that if (for some bizarre reason) a job is still in an active state (which could happen with CLI jobs) it isn't deleted after the retention period
- Make the cluster statuses be able to passed in instead of just defaulting to TERMINATED
- Remove unused property for jobs cleanup

+++ genie-web/src/main/java/com/netflix/genie/web/data/services/ApplicationPersistenceService.java
@@ -348,5 +348,5 @@ public interface ApplicationPersistenceService {
      *                         desired.
      * @return The number of successfully deleted applications
      */
-    int deleteUnusedApplicationsCreatedBefore(Instant createdThreshold);
+    long deleteUnusedApplicationsCreatedBefore(Instant createdThreshold);
 }
